+++ app/controllers/movies_controller.rb
@@ -1,6 +1,11 @@
 class MoviesController < ApplicationController
   def index
-    @movies = MoviesFacade.get_top_movies
-    require "pry"; binding.pry
+    if params[:q] == "top_rated"
+      @movies = MoviesFacade.get_top_movies
+    else params[:q] == "search"
+      # @search = params[:search]
+      @movies = MoviesFacade.get_search(params[:q])
+      # require "pry"; binding.pry
+    end
   end
 end
